api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
sqlmodel==0.0.22
pandas==2.2.2
python-dateutil==2.9.0.post0
requests==2.32.3
python-multipart==0.0.9
